+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * now `PhpDocMaker` requires the target as second argument and no longer as
     argument for the `build()` method;
 * added `PhpDocMaker::setOption()` method;
+* added `AbstractEntity::getFilename()` method;
 * improved `AbstractEntity::getDocBlockInstance()`, `AbstractEntity::getTags()`
     and `AbstractMethodEntity::getParameters()` methods. The return values are
     now cached;